cmake: Use the official CMake variable for building as a static plugin.
This is the first step in actually creating a valid Qt static plugin.

Currently, when used as a static library, it is expected to be a
subdirectory of the project instead of a "proper" Qt plugin. This
patch and the ones that follow aim to make Kirigami a plugin just
as Qt official ones.

Note that the Q_IMPORT_PLUGIN macro does *not* like the namespaced
plugin name, so *some* workaround will still be required.