generic: Get rid of most static plugin code.
A proper static plugin intended for a proper SDK isn't different
from the "normal" shared plugin. The *only* difference is the
bundling of all relevant files in the QRC.

If a `add_subdirectory(kirigami)` style subproject need to be
supported again, fine, but it should be handled differently from
the static part. It isn't the same use case and mixing subproject
and static/shared is just plain wrong.