[WIP] Add a ShadowedRectangle item
Summary:
This adds a ShadowedRectangle item, which is a rectangle with a shadow.
It uses a custom shader that draws the rectangle and the shadow below it
using distance fields, which has significantly better performance compared
to the DropShadow effect.

This is a first step to moving more of AbstractCard and related items to
C++ code. Other bits will include a ShadowedImage item and moving actual
card code to C++.

Test Plan: TBD

Reviewers: #kirigami

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D27984