introduce PassiveNotification type
Summary:
PassiveNotification is a type for small, passive and inline
notifications in the app.
used to show messages of limited importance that make sense only when
the user is using the application and wouldn't be suited as a global
system-wide notification.
This is not a full-fledged notification system. the applciation should
use this with care and only one notification should be visible at once
per app.

Test Plan:
it looks and behaves the same as ApplicationWindow.showPassiveNotification
(which now uses it) as well standalone notification items work.

Reviewers: #kirigami

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D28103